MB-67589: Fix C4244 reported by MSVC...
Warns about conversion with possible loss of data

Reviewed-on: https://review.couchbase.org/c/platform/+/237047
Tested-by: Build Bot <build@couchbase.com>

+++ cbcompress/compress_bench.cc
@@ -33,7 +33,7 @@ static void SnappyCompress(benchmark::State& state) {
         benchmark::ClobberMemory();
 #endif
     }
-    state.counters["compressed"] = compressed_length;
+    state.counters["compressed"] = static_cast<double>(compressed_length);
 }
 
 BENCHMARK(SnappyCompress)->RangeMultiplier(FACTOR)->Range(START, END);
